protected endpoints now require authorization.

+++ project1/src/main/java/com/revature/controllers/LocationController.java
@@ -11,10 +11,13 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.revature.models.Location;
+import com.revature.models.Role;
+import com.revature.services.AuthService;
 import com.revature.services.LocationService;
 
 @RestController
@@ -22,11 +25,13 @@ import com.revature.services.LocationService;
 public class LocationController {
 
 	private LocationService lServ;
+	private AuthService aServ;
 	
 	@Autowired
-	public LocationController(LocationService lServ) {
+	public LocationController(LocationService lServ, AuthService aServ) {
 		super();
 		this.lServ = lServ;
+		this.aServ = aServ;
 	}
 	
 	@GetMapping
@@ -40,19 +45,28 @@ public class LocationController {
 	}
 	
 	@PostMapping
-	public ResponseEntity<Location> postLocation(@RequestBody Location location) {
+	public ResponseEntity<Location> postLocation(@RequestHeader(value = "Authorization", required = false) String token, @RequestBody Location location) {
+		if (!aServ.checkPermission(token, Role.SELLER, Role.ADMIN)) {
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
 		Location newLocation = lServ.createLocation(location);
 		return new ResponseEntity<>(newLocation, HttpStatus.CREATED);
 	}
 	
 	@PutMapping("/{id}")
-	public ResponseEntity<Location> updateLocation(@PathVariable("id") int id, @RequestBody Location location) {
+	public ResponseEntity<Location> updateLocation(@RequestHeader(value = "Authorization", required = false) String token, @PathVariable("id") int id, @RequestBody Location location) {
+		if (!aServ.checkPermission(token, Role.SELLER, Role.ADMIN)) {
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
 		location.setId(id);
 		return new ResponseEntity<>(lServ.updateLocation(location), HttpStatus.OK);
 	}
 	
 	@DeleteMapping("/{id}")
-	public ResponseEntity<String> deleteLocation(@PathVariable("id") int id) {
+	public ResponseEntity<String> deleteLocation(@RequestHeader(value = "Authorization", required = false) String token, @PathVariable("id") int id) {
+		if (!aServ.checkPermission(token, Role.SELLER, Role.ADMIN)) {
+			return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+		}
 		lServ.deleteLocation(id);
 		return new ResponseEntity<>("Location was deleted.", HttpStatus.OK);
 	}
